refactor: create helper functions
