+++ src/archiving/decompress.js
@@ -2,27 +2,11 @@ import path from 'path';
 import { createReadStream, createWriteStream } from 'fs';
 import { pipeline } from 'stream/promises';
 import { createBrotliDecompress } from 'zlib';
+import getPaths from '../utils/getPaths.js';
 
 const decompressFile = async (consoleData, pathToHomeDir) => {
     try {
-        let pathPart = consoleData
-            .split(' ').slice(1).toString().replaceAll('"', '\'');
-
-        let archive;
-        if (pathPart.startsWith('\'')) {
-            archive = pathPart.slice(1, pathPart.indexOf('\'', 1)).replaceAll(',', ' ');
-        } else 
-            archive = consoleData.split(' ').splice(1, 1).toString();
-
-        let pathToDir;
-        if (pathPart.endsWith('\'')) {
-            const reversion = pathPart.split('').reverse().join('');
-            pathToDir = reversion
-                .slice(1, reversion.indexOf('\'', 1))
-                .split('').reverse().join('').replaceAll(',', ' ');
-        } else {
-            pathToDir = consoleData.split(' ').slice(-1).toString();
-        }
+        const { fileName: archive, pathToDir } = getPaths(consoleData);
 
         const pathToArchive = path.resolve(
             pathToHomeDir,
@@ -30,6 +14,8 @@ const decompressFile = async (consoleData, pathToHomeDir) => {
         );
 
         const fileName = archive.slice(0, archive.lastIndexOf('.'));
+        console.log('fileName: ', fileName);
+
         const pathToDestination = path.resolve(
             pathToHomeDir, 
             pathToDir,
