+++ src/fs/copyFile.js
@@ -1,27 +1,11 @@
 import path from 'path';
 import { createReadStream, createWriteStream } from 'fs';
 import { access } from 'fs/promises';
+import getPaths from '../utils/getPaths.js';
 
 const copyFile = async (consoleData, pathToHomeDir) => {
     try {
-        let pathPart = consoleData
-            .split(' ').slice(1).toString().replaceAll('"', '\'');
-
-        let fileName;
-        if (pathPart.startsWith('\'')) {
-            fileName = pathPart.slice(1, pathPart.indexOf('\'', 1)).replaceAll(',', ' ');
-        } else 
-            fileName = consoleData.split(' ').splice(1, 1).toString();
-
-        let pathToDir;
-        if (pathPart.endsWith('\'')) {
-            const reversion = pathPart.split('').reverse().join('');
-            pathToDir = reversion
-                .slice(1, reversion.indexOf('\'', 1))
-                .split('').reverse().join('').replaceAll(',', ' ');
-        } else {
-            pathToDir = consoleData.split(' ').slice(-1).toString();
-        }
+        const { fileName, pathToDir } = getPaths(consoleData);
 
         const pathToFile = path.resolve(
             pathToHomeDir,
@@ -50,10 +34,11 @@ const copyFile = async (consoleData, pathToHomeDir) => {
         );
 
         let content = '';
-        let flag = true;
         readStream.on('data', chunk => {
             content += chunk.toString();
         });
+
+        let flag = true;
         readStream.on('end', () => {
             writeStream.write(content);
         });
