+++ src/hash/hash.js
@@ -1,14 +1,11 @@
 import path from 'path';
 import { readFile } from 'fs/promises';
 import { createHash } from 'crypto';
+import getConsolePath from '../utils/getConsolePath.js';
 
 const calculateHash = async (consoleData, pathToHomeDir) => {
     try {
-        let consolePath = consoleData.split(' ').slice(1).toString();
-        if ((consolePath.startsWith('\'') && consolePath.endsWith('\''))
-            || (consolePath.startsWith('"') && consolePath.endsWith('"'))) {
-            consolePath = consolePath.replace(',', ' ').slice(1, -1);
-        }
+        const consolePath = getConsolePath(consoleData);
 
         const pathToFile = path.resolve(
             pathToHomeDir,
