+++ src/navigation/cd.js
@@ -1,14 +1,11 @@
 import path from 'path';
 import { stat } from 'fs/promises';
 import goUp from './goUp.js';
+import getConsolePath from '../utils/getConsolePath.js';
 
 const changeDir = async (consoleData, pathToHomeDir) => {
     try {
-        let consolePath = consoleData.split(' ').slice(1).toString();
-        if ((consolePath.startsWith('\'') && consolePath.endsWith('\''))
-            || (consolePath.startsWith('"') && consolePath.endsWith('"'))) {
-            consolePath = consolePath.replace(',', ' ').slice(1, -1);
-        }
+        const consolePath = getConsolePath(consoleData);
 
         const fullPathToDir = path.resolve(pathToHomeDir, consolePath);
         const isDirectory = (await stat(fullPathToDir)).isDirectory();
